requirements.txt
Pillow
pillow-heif
send2trash